src/main/java/net/ripe/commons/certification/validation/fieldvalidators/ValidationResult.java
/**
 * The BSD License
 *
 * Copyright (c) 2010, 2011 RIPE NCC
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, are permitted provided that the following conditions are met:
 *   - Redistributions of source code must retain the above copyright notice,
 *     this list of conditions and the following disclaimer.
 *   - Redistributions in binary form must reproduce the above copyright notice,
 *     this list of conditions and the following disclaimer in the documentation
 *     and/or other materials provided with the distribution.
 *   - Neither the name of the RIPE NCC nor the names of its contributors may be
 *     used to endorse or promote products derived from this software without
 *     specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
 * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
 * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
 * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
 * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */
package net.ripe.commons.certification.validation.fieldvalidators;

import java.io.Serializable;
import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;
import java.util.Set;

import net.ripe.commons.certification.validation.ValidationCheck;

import org.apache.commons.lang.Validate;
import org.apache.commons.lang.builder.ToStringBuilder;
import org.apache.commons.lang.builder.ToStringStyle;

public class ValidationResult implements Serializable {

    private static final long serialVersionUID = 1L;

    private Set<ValidationCheck> checkList =  new HashSet<ValidationCheck>();


    public boolean isTrue(boolean condition, String key, Object... param) {
        Validate.notNull(key, "key is required");
        checkList.add(new ValidationCheck(condition, key, param));
        return condition;
    }

    public boolean isFalse(boolean condition, String key, Object... param) {
        return isTrue(!condition, key, param);
    }

    public boolean notNull(Object object, String key, Object... param) {
        return isTrue(object != null, key, param);
    }

    public boolean hasFailures() {
        for (ValidationCheck vc : checkList) {
            if (!vc.isOk()) {
                return true;
            }
        }
        return false;
    }

    public List<ValidationCheck> getFailures() {
        List<ValidationCheck> failedList = new ArrayList<ValidationCheck>();
        for (ValidationCheck vc : checkList) {
            if (!vc.isOk()) {
                failedList.add(vc);
            }
        }
        return failedList;
    }

    @Override
    public String toString() {
        return ToStringBuilder.reflectionToString(this, ToStringStyle.SHORT_PREFIX_STYLE);
    }
}
